+++ lib/hitnmiss/in_memory_driver.rb
@@ -7,10 +7,22 @@ module Hitnmiss
       @cache = {}
     end
 
-    def set(key, value, expiration_in_seconds)
-      expiration = Time.now.utc.to_i + expiration_in_seconds
-      @mutex.synchronize do
-        @cache[key] = { 'value' => value, 'expiration' => expiration }
+    def set(key, entity)
+      if entity.expiration
+        expiration = Time.now.utc.to_i + entity.expiration
+        @mutex.synchronize do
+          @cache[key] = { 'value' => entity.value, 'expiration' => expiration }
+          @cache[key]['fingerprint'] = entity.fingerprint if entity.fingerprint
+          @cache[key]['updated_at'] = internal_timestamp
+          @cache[key]['last_modified'] = entity.last_modified if entity.last_modified
+        end
+      else
+        @mutex.synchronize do
+          @cache[key] = { 'value' => entity.value }
+          @cache[key]['fingerprint'] = entity.fingerprint if entity.fingerprint
+          @cache[key]['updated_at'] = internal_timestamp
+          @cache[key]['last_modified'] = entity.last_modified if entity.last_modified
+        end
       end
     end
 
@@ -20,15 +32,16 @@ module Hitnmiss
         cached_entity = @cache[key].dup if @cache[key]
       end
 
-      if cached_entity
+      return Hitnmiss::Driver::Miss.new if cached_entity.nil?
+
+      if cached_entity.has_key?('expiration')
         if Time.now.utc.to_i > cached_entity['expiration']
-          return Hitnmiss::Driver::Miss.new 
-        else
-          return Hitnmiss::Driver::Hit.new(cached_entity['value'])
+          return Hitnmiss::Driver::Miss.new
         end
-      else
-        return Hitnmiss::Driver::Miss.new
       end
+
+      return Hitnmiss::Driver::Hit.new(cached_entity['value'],
+                                       build_hit_keyword_args(cached_entity))
     end
 
     def all(keyspace)
@@ -54,8 +67,28 @@ module Hitnmiss
     end
 
     def match_keyspace?(key, keyspace)
-      regex = Regexp.new("^#{keyspace}\\" + Repository::KEY_COMPONENT_SEPARATOR)
+      regex = Regexp.new("^#{keyspace}\\" + Repository::KeyGeneration::KEY_COMPONENT_SEPARATOR)
       return regex.match(key)
+    end
+
+    private
+
+    def internal_timestamp
+      Time.now.utc.iso8601
+    end
+
+    def build_hit_keyword_args(cached_entity)
+      options = {}
+      if cached_entity.has_key?('fingerprint')
+        options[:fingerprint] = cached_entity['fingerprint']
+      end
+      if cached_entity.has_key?('updated_at')
+        options[:updated_at] = Time.parse(cached_entity['updated_at'])
+      end
+      if cached_entity.has_key?('last_modified')
+        options[:last_modified] = cached_entity['last_modified']
+      end
+      return **options
     end
   end
 end